+++ src/components/provider/provider.tsx
@@ -1,7 +1,26 @@
-import { useState, useEffect, Fragment } from "react"
-import { BlocProviderProps, BlocProviderState, BlocResolver } from "../../types"
-import { removeBlocContext } from "../../context/context"
-import { getStateFromProps } from "./util"
+import { useState, Fragment, createContext, useEffect } from "react"
+import {
+  BlocModule,
+  BlocProviderProps,
+  BlocProviderState,
+  BlocResolver,
+} from "../../types"
+import {
+  addBlocContext,
+  getBlocContext,
+  hasBlocContext,
+  removeBlocContext,
+} from "../../context/context"
+import { asClass, createContainer } from "awilix"
+
+// Don't export the root container directly
+const rootContainer = createContainer()
+
+export const getRegistrations = () => rootContainer.registrations
+
+export const registerModules = (modules: BlocModule[]) => {
+  modules.forEach((module) => module(rootContainer))
+}
 
 export function BlocProvider(
   props: React.PropsWithChildren<BlocProviderProps>,
@@ -14,8 +33,15 @@ export function BlocProvider(
     setState(stateFromProps)
 
     if (props.onCreate) {
-      const getter: BlocResolver = (blocClass) => {
-        return stateFromProps.container.resolve(blocClass)
+      const getter: BlocResolver = (blocClass, scope) => {
+        const name = scope ? `${scope}-${blocClass.name}` : blocClass.name
+        const context = getBlocContext(name)
+        if (!context) {
+          throw new Error(
+            `BlocProvider: BlocProvider could not find context "${name}"`,
+          )
+        }
+        return context.container.resolve(name)
       }
 
       props.onCreate(getter)
@@ -31,11 +57,60 @@ export function BlocProvider(
 
   if (state) {
     return (
-      <state.blocContext.Provider value={state.container}>
+      <state.blocContext.context.Provider value={state.container}>
         {props.children}
-      </state.blocContext.Provider>
+      </state.blocContext.context.Provider>
     )
   }
 
   return <Fragment></Fragment>
+}
+
+const getStateFromProps = ({
+  container,
+  bloc,
+  scope,
+}: BlocProviderProps): BlocProviderState => {
+  const providerContainer = container ?? rootContainer.createScope()
+  const shouldDestroy = !container // only destroy if the container is created by the provider
+  let names = bloc
+    .map((blocClass) => {
+      const name = scope ? `${scope}-${blocClass.name}` : blocClass.name
+      if (hasBlocContext(name))
+        throw new Error(
+          `BlocProvider.getStateFromProps: ${name} already exists in current provider context`,
+        )
+
+      return name 
+    })
+    .sort()
+    .join("-")
+
+  if (scope) {
+    names = `${scope}-${names}`
+  }
+
+  let blocContext = getBlocContext(names)
+
+  if (!blocContext) {
+    blocContext = {
+      context: createContext(providerContainer),
+      container: providerContainer,
+    }
+    blocContext.context.displayName = names
+    addBlocContext(names, blocContext)
+  }
+
+  for (let b of bloc) {
+    const name = scope ? `${scope}-${b.name}` : b.name
+    providerContainer.register({
+      [name]: asClass(b)
+        .disposer((_bloc) => _bloc.close())
+        .scoped(),
+    })
+
+    addBlocContext(name, blocContext)
+  }
+
+  return { blocContext, container: providerContainer, shouldDestroy }
 }