+++ src/hooks/use-bloc.ts
@@ -1,14 +1,14 @@
 import { BlocBase, ClassType } from "@bloc-state/bloc"
 import { useObservableState } from "observable-hooks"
-import { UseBlocSelectorConfig } from "../types"
+import { UseBlocConfig, UseBlocSelectorConfig } from "../types"
 import { useBlocInstance } from "./use-bloc-instance"
 import { useBlocSelector } from "./use-bloc-selector"
 
 export function useBloc<P, B extends BlocBase<any>>(
   bloc: ClassType<B>,
-  config?: UseBlocSelectorConfig<B, P>,
+  config?: UseBlocSelectorConfig<B, P> & UseBlocConfig,
 ): ReturnType<() => [P, B]> {
-  const providedBloc = useBlocInstance(bloc)
+  const providedBloc = useBlocInstance(bloc, config?.scope)
   if (config?.selector) {
     return [useBlocSelector(bloc, config), providedBloc]
   } else {