+++ tests/hooks/use-bloc-selector.spec.tsx
@@ -1,38 +1,29 @@
-import "reflect-metadata"
-import {
-  act,
-  cleanup,
-  render,
-  renderHook,
-  waitFor,
-} from "@testing-library/react"
+import { cleanup, renderHook } from "@testing-library/react"
 import { clearBlocContext } from "../../src/context/context"
-import { container as rootContainer } from "tsyringe"
 import { useBlocSelector } from "../../src"
-import {
-  UserBloc,
-  UserBlocProvider,
-  UserLastNameAsyncChangedEvent,
-} from "../test-helpers"
+import { UserBloc } from "../test-helpers"
 import CounterCubit from "../test-helpers/counter/counter.cubit"
 import {
   blocUserWrapper as buw,
   cubitCounterWrapper as ccw,
 } from "../test-helpers/wrappers"
+import { AwilixContainer, createContainer } from "awilix"
 
 describe("useBlocSelector", () => {
   let cubitCounterWrapper: ({ children }: any) => JSX.Element
   let blocUserWrapper: ({ children }: any) => JSX.Element
+  let container: AwilixContainer
 
   beforeEach(() => {
     cubitCounterWrapper = ccw
     blocUserWrapper = buw
+    container = createContainer()
   })
 
   afterEach(() => {
     clearBlocContext()
     cleanup()
-    rootContainer.reset()
+    container.dispose()
   })
 
   it("should return derived state for non loadable State types", () => {
@@ -60,49 +51,5 @@ describe("useBlocSelector", () => {
     )
 
     expect(result.current).toBe(0)
-  })
-
-  it("should return derived state with suspense", async () => {
-    expect.assertions(4)
-    const container = rootContainer.createChildContainer()
-    const { getByTestId } = render(UserBlocProvider(container))
-
-    await waitFor(() => {
-      const loading = getByTestId("test-loading")
-      expect(loading.innerHTML).toBe("loading")
-    })
-
-    await waitFor(
-      () => {
-        getByTestId("test-loaded")
-        expect(getByTestId("test-name").innerHTML).toBe("richards")
-      },
-      {
-        timeout: 3000,
-      },
-    )
-
-    const userBloc = container.resolve(UserBloc)
-
-    act(() => {
-      userBloc.add(new UserLastNameAsyncChangedEvent())
-    })
-
-    await waitFor(() => {
-      const loading = getByTestId("test-loading")
-      expect(loading.innerHTML).toBe("loading")
-    })
-
-    await waitFor(
-      () => {
-        getByTestId("test-loaded")
-        expect(getByTestId("test-name").innerHTML).toBe("richards")
-      },
-      {
-        timeout: 3000,
-      },
-    )
-
-    container.dispose()
   })
 })