+++ tests/provider/context.spec.tsx
@@ -1,24 +1,29 @@
-import "reflect-metadata"
 import { createContext } from "react"
-import { container } from "tsyringe"
 import {
   BlocContext,
   addBlocContext,
   getBlocContext,
   removeBlocContext,
 } from "../../src/context/context"
+import { AwilixContainer, createContainer } from "awilix"
 
 describe("BlocContextGlobalMap", () => {
   let blocContext: BlocContext
+  let container: AwilixContainer
 
   beforeEach(() => {
-    blocContext = createContext(container)
-    blocContext.displayName = "test"
+    container = createContainer()
+    const context = createContext(container)
+    context.displayName = "test"
+    blocContext = {
+      context,
+      container: container,
+    }
     addBlocContext("test", blocContext)
   })
 
   afterAll(() => {
-    container.reset()
+    container.dispose()
   })
 
   it("should remove all context from the blocContextMap", () => {