+++ tests/provider/provider.spec.tsx
@@ -1,15 +1,15 @@
-import "reflect-metadata"
 import { cleanup, render } from "@testing-library/react"
 import { UserBloc, UserMultiBlocProvider } from "../test-helpers"
 import { addBlocContext, clearBlocContext } from "../../src/context/context"
 import { createContext } from "react"
-import { container as parentContainer } from "tsyringe"
+import { AwilixContainer, createContainer } from "awilix"
 
 describe("BlocProvider", () => {
   const originalConsoleError = console.error
+  let container: AwilixContainer
 
   beforeEach(() => {
-    parentContainer.reset()
+    container = createContainer()
     console.error = () => {}
   })
 
@@ -17,12 +17,16 @@ describe("BlocProvider", () => {
     console.error = originalConsoleError
     cleanup()
     clearBlocContext()
+    container.dispose()
   })
 
   it("should throw an error if a provided bloc already exists in contextMap", async () => {
-    const blocContext = createContext(parentContainer)
+    const blocContext = createContext(container)
     blocContext.displayName = UserBloc.name
-    addBlocContext(UserBloc.name, blocContext)
+    addBlocContext(UserBloc.name, {
+      context: blocContext,
+      container: container,
+    })
 
     expect(() => {
       render(UserMultiBlocProvider())