+++ tests/test-helpers/user/components/user-bloc.provider.tsx
@@ -1,78 +1,49 @@
+import { AwilixContainer } from "awilix"
 import { Suspense } from "react"
-import { DependencyContainer } from "tsyringe"
-import {
-  UserBloc,
-  UserBlocListenerEvent,
-  UserLastNameAsyncChangedEvent,
-  UserMultiBlocListenerEvent,
-} from ".."
+import { UserBloc, UserLastNameAsyncChangedEvent } from "../user"
 import { BlocProvider } from "../../../../src"
 import { BlocListener } from "../../../../src/components/bloc-listener/listener"
-import CounterCubit from "../../counter/counter.cubit"
 import { UserBlocConsumer } from "./user-bloc-consumer"
+import CounterCubit from "../../counter/counter.cubit"
 
-export const UserMultiBlocListenerProvider = (container?: DependencyContainer) => (
+export const UserSingleBlocListenerProvider = (container?: AwilixContainer) => (
   <BlocProvider
-    bloc={[UserBloc, CounterCubit]}
-    name="user-multi-bloc-provider"
+    bloc={[UserBloc]}
     container={container}
-    onCreate={(get) => get(UserBloc).add(new UserLastNameAsyncChangedEvent())}
-  >
-    <BlocListener
-      bloc={[UserBloc, CounterCubit]}
-      listen={(get, state) => {
-        const count = state as number
-        get(UserBloc).add(new UserMultiBlocListenerEvent(count))
-      }}
-      listenWhen={ ( get, state ) => {
-        return typeof state === "number"
-      }}
-    >
-      <UserBlocConsumer />
-      </BlocListener>
-  </BlocProvider>
-)
-
-export const UserSingleBlocListenerProvider = (
-  container?: DependencyContainer,
-) => (
-  <BlocProvider
-    bloc={UserBloc}
-    container={container}
-    onCreate={(get) => get(UserBloc).add(new UserLastNameAsyncChangedEvent())}
+    onCreate={(get) => {
+      get(UserBloc).add(new UserLastNameAsyncChangedEvent("richards"))
+    }}
   >
     <BlocListener
       bloc={UserBloc}
-      listen={(get, state) => {
-        get(UserBloc).add(new UserBlocListenerEvent())
+      listener={(get, state) => {
+        get(UserBloc).add(new UserLastNameAsyncChangedEvent("bloc-listener"))
       }}
     >
-      <UserBlocConsumer swr={false} suspend={false} />
+      <UserBlocConsumer />
     </BlocListener>
   </BlocProvider>
 )
 
-export const UserBlocProvider = (
-  container?: DependencyContainer,
-  swr?: boolean,
-) => (
+export const UserBlocProvider = (container?: AwilixContainer) => (
   <BlocProvider
-    bloc={UserBloc}
+    bloc={[UserBloc]}
     container={container}
-    onCreate={(get) => get(UserBloc).add(new UserLastNameAsyncChangedEvent())}
+    onCreate={(get) =>
+      get(UserBloc).add(new UserLastNameAsyncChangedEvent("richards"))
+    }
   >
     <Suspense fallback={<div data-testid="test-loading">loading</div>}>
-      <UserBlocConsumer swr={swr} />
+      <UserBlocConsumer />
     </Suspense>
   </BlocProvider>
 )
 
-export const UserMultiBlocProvider = (container?: DependencyContainer) => (
+export const UserMultiBlocProvider = (container?: AwilixContainer) => (
   <BlocProvider
     bloc={[UserBloc, CounterCubit]}
-    name="user-multi-bloc-provider"
     container={container}
-    onCreate={(get) => get(UserBloc).add(new UserLastNameAsyncChangedEvent())}
+    onCreate={(get) => get(UserBloc).add(new UserLastNameAsyncChangedEvent(""))}
   >
     <Suspense fallback={<div data-testid="test-loading">loading</div>}>
       <UserBlocConsumer />